==> 3. Delete NoCalls & Span.ipynb <==
from sklearn.svm import LinearSVC
from sklearn import preprocessing
import numpy as np
import pandas as pd
from sklearn.model_selection import cross_val_score, train_test_split
from sklearn.metrics import accuracy_score, confusion_matrix
import matplotlib.pyplot as plt
%matplotlib inline
import collections
import os
import seaborn
seaborn.set()

encoded = np.load('./npy_data/encoded.npy')
varvals = np.load('./npy_data/varvals.npy')
pathdataOH = np.load("./npy_data/path_data_oh.npy")
oldpath = np.load('./npy_data/old_path.npy')

# deleting the missing and spanning tiles - keep everything that is >= 2 (delete 0 and 1)
to_keep = varvals >= 2 
# get the indices
idxTK = np.nonzero(to_keep)
idxTK = idxTK[0]

encoded.shape

encoded = encoded[:, idxTK]

varvals = varvals[idxTK]
pathdataOH = pathdataOH[idxTK]
oldpath = oldpath[idxTK]

np.save("./npy_data/varvals_d.npy", varvals)
np.save("./npy_data/path_data_oh_d.npy", pathdataOH)
np.save("./npy_data/old_path_d.npy", oldpath)

np.save("./npy_data/data_encoded_d.npy", encoded)
