==> 5. Read Coefficients.ipynb <==
from sklearn.svm import LinearSVC
from sklearn import preprocessing
import numpy as np
import pandas as pd
from sklearn.model_selection import cross_val_score, train_test_split
from sklearn.metrics import accuracy_score, confusion_matrix
import matplotlib.pyplot as plt
%matplotlib inline
import collections
import os
import seaborn
seaborn.set()
from sklearn.externals import joblib

svc = joblib.load('svc.pkl')
encoded = np.load('./npy_data/data_encoded_d.npy')
varvals = np.load('./npy_data/varvals_d.npy')
pathdataOH = np.load("./npy_data/path_data_oh_d.npy")
oldpath = np.load('./npy_data/old_path_d.npy')

varvals

varvals[20431281]

oldpath[13933181]

# retrieve all the nonzero coefficients and zip them with their respective indices
nonzeroes = np.nonzero(svc.coef_[0])[0]
coefs = zip(nonzeroes, svc.coef_[0][nonzeroes])

# sort the coefficients by their value, instead of index
coefs.sort(key = lambda x: x[1], reverse=True)

for coef in coefs:
    print coef
    print "Old path: ", oldpath[coef[0]]

maxCoef = np.absolute(svc.coef_).max()

idxM = np.argmax(np.absolute(svc.coef_))
numnz = np.nonzero(svc.coef_)[1].shape
idxNZus = np.nonzero(svc.coef_)[1]
coefs = svc.coef_[0,:]

nnzcoefs = coefs[idxNZus]

idxSort = np.argsort(np.absolute(nnzcoefs))
idxSort = np.flipud(idxSort)
idxNZ = idxNZus[idxSort]

coefPaths = pathdataOH[idxNZ]

tile_path = np.trunc(coefPaths/(16**5))
tile_step = np.trunc((coefPaths - tile_path*16**5)/2)
tile_phase = np.trunc((coefPaths- tile_path*16**5 - 2*tile_step))

print(tile_path, tile_step, tile_phase)

vhex = np.vectorize(hex)

vectorizedPath = vhex(tile_path.astype('int'))
vectorizedStep = vhex(tile_step.astype('int'))
vectorizedPhase = vhex(tile_phase.astype('int'))

tile_loc = np.column_stack((tile_path, tile_step))
print(tile_loc)
print(nnzcoefs[idxSort])
print(oldpath[idxNZ])
print(varvals[idxNZ])

def tileSearch(arg):
    vecpath = str(vectorizedPath[int(arg)])
    vecpath = vecpath[2:].zfill(4)
    try:
        proc = subprocess.check_output("cat " + args.assembly_fwi + " | grep :" + vecpath, shell=True)
        return proc
    except CalledProcessError as e:
        return "Assembly index file not found or `cat` command not available. Continuing..."
    

tileSearch(12559433)
